test: add proper search screen snapshot test
